requirements.txt
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
XlsxWriter==3.1.9
